ekf2: add kconfig for magnetometer support (enabled by default)
